added a new function in the parser
This is the first commit on the rework of the parser for optimization pupose. Now we parse the serialized string before calling the deserialize method
of the AttrContainer. AttrContainer::deserialize now takes a map<std::string, std::string> as argument. This map contains the ids of the attributes
and there values. This method should be faster as the previous one as we do not give the entire serialized string to elements during the deserialization
part. By doing this, we reduce the number of times each part of the string is read by the deserializer.

Notice that this method is still less efficient than a real parser as we are still reading characters multiple times (especially when there is a lot of
nested objects).

This change also makes the parser safer as well. Indeed, we used to have a find in deserialize to search for the ids of the attributes in the entire
serialized string. This could lead to a problem if the same id was used for attributes in nested objects. For instance:
"{ p: { x: 0, y: 0 }, x: 1 }" (the find catches the first x, which is equal to 0, instead of the one that is eaqual to 1)