support for std::set + reorganize convertor functions
The convertor functions for strings were defined after those for containers. This was a probleme with the test on the set.
note: This wasn't tested properly as we had no test with string containers before (the map stores a pair and not directly a string).
todo: analyse and document dependancies between function (and update the tests).