fix: nested containers were not supported
The problem was that we were checking that the contained type was
serializable in the template of the function responsible of serializing
containers. The compiler was searching for the function that could
handle containers but this one wasn't technically treated yet as the
template wasn't fully read.

The problematic check has been removed from the template, however this
is a problem for debuging as this check was there to help the user to
understand that the contained type is not supported.

todo: is is very important to add error message using static assert.