start changing the string format and the way data is encode
- All the unecessary character / separators should be removed: we want
to optimize the size of the output as well as the serialization /
parsing time.
- now the value are binary encoded using reinterpret_cast: the output
size is smaller that a regular string and the conversion is faster
compared to using a stream.
- stream are slow: they shouldn't be used.

todo: the way custom convertors are defined currently won't work with
the new version of the library: this should be redesigned.