first optimization update
- now we use string and reinterpret_cast to serialize the elements. The binary format is liter and values can be added to the string without additional separators. This makes the parser faster as well.
- all the tests pass for this version.

todo: since the order of the elements will never change, we don't need the ids of the attributes, so they should be removed from the AttrContainer.
todo: macros for generating custom convertor are not that practical anymore. They should be changed (the string serialize / deserialize functions from the convertor might be handy here, so it would be nice to be able to access them from the custom convertor functions).
todo: now that the optimization work has been started, timed tests / timed program would be nice to make measures.
todo: the documentation should be updated (as well as some old comments maybe).