convertor change
Originally, the lib used a big macro containing serialize and deserialize function to allow the user to add support for external types. This solution was found because we couldn't use virtual functions since all the conversion functions where templates.
This previous solution was bad as a lot of code was generated with the macro, and it was terrible to maintain.
The new solution uses multiple inheritance. The default convertor (that contains the old serialize template functions) inherits from various Convert classes (one by new type), and each of these has pure virtual conversion functions that must be overridden in the user's convertor class. The custom convertor doesn't require the usage of the CONVERTOR macro anymore. Instead, it must inherit from the Convertor and specify the types for which it adds support for in the template parameter list.

The deserialize_custom_type and serialize_custom_type macros have been removed and the DESERIALIZE_POLYMORPHIC macros has been replaced by HANDLE_POLYMORPHIC that overrides the serialize function as well as the deserialize function (both have to be writen because they are pure virtual in the base class).

The test convertors have been updated as well and all the test pass.

todo: test with QString