refactor: rename AttrContainer into MemberList
The term AttrContainer wasn't properly writen and was too long. Further, using the word "container" was not precise enough since the data structure is actually a list (with nodes like a linked list but with a size known at compile time).
The term "Member" is preferred to "Attribute" since it suits better to C++ vocabulary. Moreover, this term describes the variables as well as the functions, and it might be useful to store functions in the member list in the future. Those won't be serialized (there is no point of doing such a thing in our case) but they could be executed to change the data at any point of the serialization.

note: we could have used the term "table" instead of "list" to describe the data structure since it stores the identifiers of the attributes. However, the term table should be used to refer to a data structure that is made to be looked up, unlike our list. Indeed, we do not use the identifier for a search purpose primarily. This could be done in the case of a deserialization of a JSON file (not implemented at this stage) but this wouldn't be as efficient as a real table.