perfomance test
A performance test has been added to compare the speed of the serializer with the speed of c serialization (cast of a basic struct).

For now, the serialization appears to be really slow and should be optimized. The deserialization is slightly faster since we don't have the manage the output string (we may have some mallocs to resize the string array in the serialization).