fix: polymorphism not supported with smart pointers
The macro for generating the functions that can serialize polymorphic types was generating the code only for standard pointers. Now, shared and unique pointers work as well.

fix: explicit constructor could cause ambiguous function call with custom convertors.

Now we use a static_cast to make sure to call the correct function from the Convert mother class. Otherwise, we could have an error if some constructor where no explicit (ex: pointers and shared_ptrs).

fix: impossible to deserialize an array of unique_ptr because of the copy in the insert functions.

The test have been updated to make sure this new cases are verified.