function execution during the serialization
Now we can pass functions to the serializer. These functions can be
executed at anytime depending on the position of the function in the
member list. Here functions can be used to execute some code during the
serialization or the deserialisation.

Function type: std::function<void(Phases, std::string_view const&)>
We can use the phase to know if we are serializing or deserializing.

todo: such a feature should be used to link pointers. A macro should be
writen to do this easily (try to serialize a double linked list or a
tree).