Start adding support for dynamic arrays.
A wrapper class has been added to manage the pointer and its dimensions.
We can have any dimensions and the wrapper class constructor can be
configured to take references on the size as well as values.

note: The pointers are managed may change in the future. For now, the
user should make sure that either the pointer is null, or it is properly
allocated with the correct size. This may lead to issues if the pointers
are not managed correctly.

todo: The current organization and naming for the file should be
improved.
todo: test pointer link macro.