csstruct optimization
A new special type CStruct has been added. Similarly to DynamicArray, it
is a wrapper type that can encapsulate a reference to a cstruct type.

note: there is no way to verify if the given type is a cstruct or not so
it's up to the user.

todo: the default serialization is slow.