New interface (#3)
* start adding constexpr in the convertor for future optimizations

* changed the concepts for serializable and deserializable types

* start working on the new interface

* update some of the tests

problem: const and dynamic arrays

* compiling but the resize doesn't work

* start implementing the new convertor

Now the memory is a member of the convertor class. The objective is to
reduce the stack usage with multiple nested serialize functions call.

The output type is customizable threw a template parameter (the idea is
to use something faster than the strings)

* pass composed test

note: the size of the output is not perfect at the end.

* more tests

todo: find a way to resize the memory properly.

* introducing a custom vector for the output

The vector from the standard library does not have all the methods I
need. The tools available in std are slower than the tailored data
structure.

* fix: missing test in vec + pass more tests

* use forwarding references

* passing test on maps

The const problem was solve in a better way compared to the previous
version. In fact, removing the const from the pair works well.

todo: it's quite messy in the metafunctions, we should have another file
that could be used to put metafunctions that requires the concepts.

* some fix for tuples and enums + refactor test

Test on tuples, map, set and enums pass.
Now we can use ifdef to toogle the tests.

* test on shared pointer passes

* vector convertion function for vec

* static arrays now serializable

* dynamic arrays + automatic serialization for trivialy copyable types

* remove forward in the insert functions in tools

* added an optimization for std::vector

problem: if we desize to serialize using std::vector and a non 0
position, the vector will not be resized properly

* rework serializer functions for the new interface

- Now the phase can be know compile time (fix the issue of const
    serialize function)
- The new function type take the convertor as parameter.
- A context type has been added for template deduction of the lambda
  (template parameter specification does not work after passing the
   function as auto parameter).

bug: dynamic arrays are not recognize by the convertor function anymore
todo: test the new serializer type and see if it is fast enough

* fix dynamic array + trivial types

- dynamic arrays are now serializable: the DynamicArray class has been
  changed as well as the convertor functions
- now we do not treat fundamental types anymore. Instead we treat
  trivial types which are all the types directly serializable.

* fix missing clear for containers

todo: might not be the best way to do this (if the container needs to be
    cleared but does not have a clear method).

* fix: make sure std::arrays are not treated as a trivial type

* make sure the trivial types are not serializable

* new convertor

bug: cannot deserialize using a const container

* fix: missing test for resize

* optimization for iterables of trivials (first really good performance for serialization)

* trivial optimization for static arrays

* more optimization for vectors (resize first)

* remove some useless files

* remove the types file and create a dedicated header for the context type.

* start reorganizing the meta functions and the concepts

* clean the convertor + new macros

* fix: wrong deserialization for pointers (test on the tree passes)

* update cmake

* fix: missing files in git

* first working test for the polymorphism with the new interface (type ids)

todo: id_table usefull ?
todo: rewrite the dedicated macro for polymophic types

* remove cout

* inheritance and polymorphism supported again

- the old macros have been updated
- helper functions added in the convertor and serialize
- helper functions for id table

todo: fix the concepts Serialize and Deserialize
todo: hh example

* fix: Serializer and Deserialize concepts

Now these concepts works for any memory type

* abstract test

- added more requirements for the serialize functions of the shared
  pointers (the underlying type should not be in the list of additional
  types).
- fix: id - 1 requires a cast in the idTable functions
- the macro SERIALIZE_ABSTRACT is working

* other polymorphic test

* hh test + valgrind check

* changed the hedgehog example (more realistic)

* more logic task manager

* improved the type table

* better network and task transfer

* remove some useless elements

* update comments in the convertor

* comments in exceptions + remove dead code

* comments in serialize

* fix: rename functions

* context is now in the tools namespace

* comments in tools

* rename file + concept for identifier types

* IdTable -> TypeTable

* bind interface

* vec -> Bytes + comments

* comments

* move super in its own file

* move the getId function in TypeTable

* fix: missing constexpr inline specifier

* deserializeSize is now template

* more comments in type_table

* regroup the macros in a seprated file

* clean convertor + reorganized concepts and metafunctions

* convertor renamed into serializer

* improve the bind interface

Now we can use sandard getters and setters with the bind interface

* fix: comparison between char and byte could cause errors

* type alias for byte type in Bytes

* serializer include file

This file can be used to inlucde all the required serializer files. It
also contains aliases.

Note: the macros file have to be inlucded manually

* Generic update

Now ids are handled automatically by default, and there is no need to
create a custom serializer for generic types.

* test missing edge case for tuples

* the createGeneric function has been renamed since non generic types are
also created using this function

* rename SERIALIZE_CONV to SERIALIZE_CUSTOM

* improve unknown serializer

* fix tuple deserialization

* fix: mem leak in tuple test

* refactor: move functions out of macros

* SER_MemT macro

* update macros

* change getId function

* static assert tests

* refactor: createId

The createGeneric function as been remove and replaced with a createId
one that can be used to create non polymorphic objects. Exceptions have
also been added when the id is not found in the type table or when the
id corresponds to a generic type.

* update the readme

* fix cmake list

* new constructor for Bytes

* fix: the bytes buffer can now be const

todo: find a better solution

* new helper variadic functions in the serializer

* string deserialization optimization + remove old useless operations

* move some functions

* memcheck test (todo: test if it works)

* doxygen documentation

* update doc

* removed useless code

* fix pipeline